Put energy metering into the evaluation
Add a new field to contain energy metric data in the Evaluation.
Use another set of mpsc channels to communicate values back to the
main testbed thread.

Switch up the way to decide what meters to collect data for. The
meters mentioned in test criteria are allocated in the previously
cleared output vector. When sampling energy meters in the test
loop, only loop over the meters defined in the output sample
vector.