Show detailed trace information
When displaying Evaluation, show informamtion about traces seen
and their timing. There may be times where there are inaccuracies
with what trace points are encountered when. When that happens,
the values may be outside of range or show up as other points.

When they are outside of range, show UNTRACKED?