Implement basic reset functionality
Make the testbed perform a primitive reset on the device under test
when the Test specifies that the device must be reset prior to
starting the test. This only works for a POR where the reset pin
must be held low and then allowed to go high after a short period
of time.

More sophisticated resets would need to be defined by the device
under test.