Split up reset operation
Define and use two parts to the reset process instead of doing it
all at once with a single function and call. This lets us release
the DUT from reset when we are ready.

Wait for a small amount of time after placing the device in reset
as there is always some minimum amount of time that we must hold the
pin before releasing it for the signal to cause a reset in the DUT.
Use 10ms for now. This should be large enough to make a traceable
difference on a cheap oscilloscope and to cause a reset in almost
any case (our DUT requires only 10ns) yet small enough to be barely
noticable.