Move evaluation logic into clockwise_common
Last major code move of logic from the binary to the common library
of code. This allows for a test suite to define its own evaluation
logic. Now, when the Testbed executes Tests, it produces an
Observation for each Test run.

Each Observation should pass through an Evaluator which will
produce an Evaluation that shows whether the Criteria were met or
not.